Review changes

+++ internal/app/consumer/db.go
@@ -15,7 +15,7 @@ type Consumer interface {
 }
 
 type consumer struct {
-	n      uint64
+	n      int
 	events chan<- streaming.LikeEvent
 
 	repo repo.EventRepo
@@ -30,7 +30,7 @@ type consumer struct {
 }
 
 type Config struct {
-	n         uint64
+	n         int
 	events    chan<- streaming.LikeEvent
 	repo      repo.EventRepo
 	batchSize uint64
@@ -38,7 +38,7 @@ type Config struct {
 }
 
 func NewDbConsumer(
-	n uint64,
+	n int,
 	batchSize uint64,
 	consumeTimeout time.Duration,
 	repo repo.EventRepo,
@@ -64,7 +64,7 @@ func (c *consumer) Start(ctx context.Context) {
 	childContext, stopFunc := context.WithCancel(ctx)
 	c.cancel = stopFunc
 
-	for i := uint64(0); i < c.n; i++ {
+	for i := 0; i < c.n; i++ {
 		c.wg.Add(1)
 
 		go func() {
